working on removing macros
To tidy code, I plan to replace the macros with functions, and use
function pointers to specify which version of the function is to be used
for the dimension. I hoped that this might also allow me to use function
pointers to select a version of the propagation function that stored the
second derivative of the wavefield for use during the imaging condition,
however the current imaging condition also requires the first derivative
and the wavefield itself.
[ci skip]